src/Shr3y4sm-meme_generator/requirements.txt
Pillow>=10.0.0
tk>=0.1.0